Remove PatchCandidate interface
