+++ src/test/java/jp/posl/jprophet/PatchedProjectGeneratorTest.java
@@ -5,7 +5,7 @@ import org.junit.Before;
 import org.junit.Test;
 
 import jp.posl.jprophet.operation.AstOperation;
-import jp.posl.jprophet.patch.DefaultPatchCandidate;
+import jp.posl.jprophet.patch.PatchCandidate;
 import jp.posl.jprophet.patch.PatchCandidate;
 import jp.posl.jprophet.project.GradleProject;
 
@@ -71,7 +71,7 @@ public class PatchedProjectGeneratorTest{
         Node node = NodeUtility.deepCopyByReparse(targetNodeBeforeFix);
         node.getTokenRange().orElseThrow().getBegin().replaceToken(new JavaToken(node.getTokenRange().orElseThrow().getBegin().getRange().get(), JavaToken.Kind.PRIVATE.getKind(), "private", null, null));
         CompilationUnit cu = NodeUtility.reparseCompilationUnit(node.findCompilationUnit().get()).orElseThrow();
-        PatchCandidate patchCandidate = new DefaultPatchCandidate(targetNodeBeforeFix, cu, this.targetFilePath, this.targetFileFqn, AstOperation.class, 1);
+        PatchCandidate patchCandidate = new PatchCandidate(targetNodeBeforeFix, cu, this.targetFilePath, this.targetFileFqn, AstOperation.class, 1);
         this.patchedProjectGenerator.applyPatch(patchCandidate);
        
     }
