+++ src/test/java/jp/posl/jprophet/patch/PatchCandidateTest.java
@@ -17,7 +17,7 @@ import org.junit.Test;
 import jp.posl.jprophet.operation.VariableReplacementOperation;
 
 
-public class DefaultPatchCandidateTest {
+public class PatchCandidateTest {
     private PatchCandidate patchCandidate;
     private String filePath = "src/test/resources/test01.java";
     private CompilationUnit compilationUnit;
@@ -41,7 +41,7 @@ public class DefaultPatchCandidateTest {
         fixedNode.getTokenRange().orElseThrow().getBegin().replaceToken(new JavaToken(node.getTokenRange().orElseThrow().getBegin().getRange().get(), JavaToken.Kind.PRIVATE.getKind(), "private", null, null));
         LexicalPreservingPrinter.setup(fixedCompilationUnit);
         this.newFixedCompilationUnit = JavaParser.parse(LexicalPreservingPrinter.print(fixedCompilationUnit));
-        this.patchCandidate = new DefaultPatchCandidate(node, this.newFixedCompilationUnit, filePath, fqn, VariableReplacementOperation.class, 1);
+        this.patchCandidate = new PatchCandidate(node, this.newFixedCompilationUnit, filePath, fqn, VariableReplacementOperation.class, 1);
     }
 
     /**
